service/s3/s3manager: Removed unneed casting for upload part size

+++ service/s3/s3manager/upload.go
@@ -322,7 +322,7 @@ func (u *uploader) initSize() {
 
 		// Try to adjust partSize if it is too small and account for
 		// integer division truncation.
-		if float64(u.totalSize)/float64(u.opts.PartSize) >= float64(MaxUploadParts) {
+		if u.totalSize/u.opts.PartSize >= int64(MaxUploadParts) {
 			// Add one to the part size to account for remainders
 			// during the size calculation. e.g odd number of bytes.
 			u.opts.PartSize = (u.totalSize / int64(MaxUploadParts)) + 1
